==> 4_imbal_base_no_opt_t_1_2.ipynb <==
# Phase 4: Data Balancing for $t=2$ **without** Threshold Opt.
* Models: SVM, MLP, LightGBM, XGBoost, CatBoost

## Setup

# Python ≥3.5 is required
import sys
assert sys.version_info >= (3, 5)

# Common imports
import numpy as np
import pandas as pd

from imblearn.over_sampling import (
    SMOTE, ADASYN
)
from imblearn.under_sampling import (
    CondensedNearestNeighbour,
    EditedNearestNeighbours,
    TomekLinks,
)
from imblearn.combine import SMOTEENN, SMOTETomek
from imblearn.pipeline import Pipeline
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from sklearn.svm import SVC
from sklearn.neural_network import MLPClassifier
from catboost import CatBoostClassifier


## Data Loading

seed = 42
df = pd.read_csv(f"datasets/train_set_{seed}_t_2.csv")

Separate features and labels from the train-set and valid-set

X_80 = df.drop(['is_drop'], axis=1)
y_80 = df['is_drop'].copy()

X_80.info()

## Data Transformation

# Label encoding for categorical features
def sex_mapping(sex):
    sex_map = {'M': 0, 'F': 1}
    return sex_map[sex]

def seme_mapping(seme):
    seme_map = {1: 0, 2: 1}
    return seme_map[seme]

def college_mapping(college):
    college_map = {'TH': 0, 'HS': 1, 'BZ': 2, 'HT': 3, 'EG': 4, 'SW': 5, 'AT': 6}
    return college_map[college]


# Label encoding for categrical features
def label_encoding(data):
    data['sex'] = data['sex'].map(sex_mapping)
    data['semester'] = data['semester'].map(seme_mapping)
    data['college'] = data['college'].map(college_mapping)

from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer

num_attrs = ['grade', 'gpa_last_seme', 'credits_last_seme', 'credits_tot', 'n_seme', 'years_since']
cat_attrs = ['semester', 'sex', 'adm_unit', 'nation', 'in_capa', 'college', 'leave']

full_pipeline = ColumnTransformer([
    ('num', StandardScaler(), num_attrs),
    ('cat', OneHotEncoder(), cat_attrs)
])

cat_pipeline = ColumnTransformer([
    ('num', StandardScaler(), num_attrs),
    ('cat', 'passthrough', cat_attrs)
])

X_t = full_pipeline.fit_transform(X_80)    # for SVM and MLP with one-hot encoding
label_encoding(X_80)   # for XGBoost and LightGBM with label encoding
X_cat = cat_pipeline.fit_transform(X_80) # for CatBoost

X_80 = pd.DataFrame(X_t, columns=full_pipeline.get_feature_names_out())   # for SVM and MLP
X_cat = pd.DataFrame(X_cat, columns=cat_pipeline.get_feature_names_out()) # for CatBoost, LightGBM, and XGBoost
cat_indices = [6, 7, 8, 9, 10, 11, 12]

# Typesetting
num_cols = ['num__grade', 'num__gpa_last_seme', 'num__credits_last_seme', 'num__credits_tot', 'num__n_seme', 'num__years_since']
cat_cols = list(set(X_80.columns) - set(num_cols))
X_80[num_cols] = X_80[num_cols].astype(float)
X_80[cat_cols] = X_80[cat_cols].astype(int)

cat_cols = list(set(X_cat.columns) - set(num_cols))
X_cat[num_cols] = X_cat[num_cols].astype(float)
X_cat[cat_cols] = X_cat[cat_cols].astype(int)

## Model Training with Resampling

sampler_map = {
    "smote": SMOTE(random_state=0),
    "adasyn": ADASYN(
        sampling_strategy="auto",  # samples only the minority class
        random_state=0,  # for reproducibility
        n_neighbors=5,
    ),
    "smoteenn": SMOTEENN(random_state=0),
    "cnn": CondensedNearestNeighbour(
        sampling_strategy="auto", random_state=0, n_neighbors=1, n_jobs=4
    ),
    "enn": EditedNearestNeighbours(
        sampling_strategy="auto", n_neighbors=3, kind_sel="all", n_jobs=4
    ),
    "tomek": TomekLinks(sampling_strategy="auto", n_jobs=4),
}

Resampling and training

svm = {}
mlp = {}
xgb = {}
lgb = {}
cat = {}

for sampler in sampler_map.keys():
    print(sampler)

    # resample
    X, y = sampler_map[sampler].fit_resample(X_80, y_80)

    # set up the classifier
    svm[sampler] = SVC(C=10, gamma=0.1)
    svm[sampler].fit(X, y)

    mlp[sampler] = MLPClassifier(max_iter=10_000, hidden_layer_sizes=[10,10,10], random_state=43)
    mlp[sampler].fit(X, y)

    X, y = sampler_map[sampler].fit_resample(X_cat, y_80)
    
    xgb[sampler] = XGBClassifier(n_estimators=40, max_depth=5)
    xgb[sampler].fit(X, y)

    lgb[sampler] = LGBMClassifier(n_estimators=90, max_depth=8)
    lgb[sampler].fit(X, y)

    cat[sampler] = CatBoostClassifier(cat_features=cat_indices, max_depth=6, iterations=1000)
    cat[sampler].fit(X, y)

import joblib

names = ['svm', 'mlp', 'xgb', 'lgb', 'cat']
for name in names:
    model_dict = {
        'model': globals()[f'{name}'],
    }
    joblib.dump(model_dict, f'models/imbal_default_{name}_1_2.pkl')


## Final Test

test_set = pd.read_csv(f"datasets/test_set_{seed}_t_2.csv")

# Preprocessing test set
X_te = full_pipeline.transform(test_set)
label_encoding(test_set)
X_te_cat = cat_pipeline.transform(test_set)
y_te = test_set['is_drop'].copy()

X_te = pd.DataFrame(X_te, columns=full_pipeline.get_feature_names_out())
X_te_cat = pd.DataFrame(X_te_cat, columns=cat_pipeline.get_feature_names_out())

# Typesetting 
X_te_cat[num_cols] = X_te_cat[num_cols].astype(float)
X_te_cat[cat_cols] = X_te_cat[cat_cols].astype(int)

from sklearn.metrics import (
    roc_auc_score, average_precision_score,
    recall_score, precision_score, f1_score, matthews_corrcoef
)
from pprint import pprint

# function to evaluate the performance with test set and given threshold
def test_model(clf, X_test, y_test): 
    if (type(clf).__name__ == 'SVC'):
        y_score = clf.decision_function(X_test)
    else: y_score = clf.predict_proba(X_test)[:, 1]

    auc = roc_auc_score(y_test, y_score)
    ap = average_precision_score(y_test, y_score)

    if (type(clf).__name__ == 'SVC'):
        y_pred = (y_score >= 0)
    else: y_pred = (y_score >= 0.5)
    recall_no = recall_score(y_test, y_pred)
    preci_no = precision_score(y_test, y_pred)
    f1_no = f1_score(y_test, y_pred)
    mcc_no = matthews_corrcoef(y_test, y_pred)

    res = {'auc': auc, 'ap': ap, 'recall': recall_no, 'precision': preci_no, 'f1': f1_no, 'mcc': mcc_no}
    print("model: ", type(clf).__name__)
    pprint(res)
    return res

# Now we test the six models with the test set and given thresholds.

test_svm = {}
test_mlp = {}
test_xgb = {}
test_lgb = {}
test_cat = {}

for sampler in sampler_map.keys():
    print(sampler)
    
    # set up the classifier
    test_svm[sampler] = test_model(svm[sampler], X_te, y_te)
    test_mlp[sampler] = test_model(mlp[sampler], X_te, y_te)
    test_lgb[sampler] = test_model(lgb[sampler], X_te_cat, y_te)
    test_xgb[sampler] = test_model(xgb[sampler], X_te_cat, y_te)
    test_cat[sampler] = test_model(cat[sampler], X_te_cat, y_te)


    print()
